Module 3 finished late
The click event and color change still work on table. The  geojson
information appears to print out correctly. Turned in a day late.
Decided to do a new commit as the kml file was not included in the
previous commit was unsure if it was worth any points or not and if only
what contained in the commit will be looked at I didnt want to be docked
more than I already will be for it not being in the commit.